+++ app/src/main/java/org/torproject/android/OrbotMainActivity.java
@@ -23,13 +23,13 @@ import org.torproject.android.service.TorServiceConstants;
 import org.torproject.android.service.util.TorServiceUtils;
 import org.torproject.android.settings.SettingsPreferences;
 import org.torproject.android.ui.AppManager;
-import org.torproject.android.ui.hs.HiddenServicesActivity;
+import org.torproject.android.ui.hiddenservices.HiddenServicesActivity;
 import org.torproject.android.ui.ImageProgressView;
 import org.torproject.android.ui.PromoAppsActivity;
 import org.torproject.android.ui.Rotate3dAnimation;
-import org.torproject.android.ui.hs.providers.HSContentProvider;
+import org.torproject.android.ui.hiddenservices.providers.HSContentProvider;
 import org.torproject.android.vpn.VPNEnableActivity;
-import org.torproject.android.backup.BackupUtils;
+import org.torproject.android.ui.hiddenservices.backup.BackupUtils;
 
 import android.annotation.SuppressLint;
 import android.app.ActivityManager;
