+++ app/src/main/java/org/torproject/android/ui/hiddenservices/storage/ExternalStorage.java
@@ -1,4 +1,4 @@
-package org.torproject.android.storage;
+package org.torproject.android.ui.hiddenservices.storage;
 
 import android.os.Environment;
