+++ app/src/main/res/layout/layout_hs_list_view.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.torproject.android.ui.hs.HiddenServicesActivity">
+    tools:context="org.torproject.android.ui.hiddenservices.HiddenServicesActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
