+++ app/src/main/res/layout/layout_hs_list_view_main.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="org.torproject.android.ui.hs.HiddenServicesActivity"
+    tools:context="org.torproject.android.ui.hiddenservices.HiddenServicesActivity"
     tools:showIn="@layout/layout_hs_list_view">
 
     <ListView
